fix: 修改删除条件时条件判断

+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcAdServiceImpl.java
@@ -52,7 +52,9 @@ public class DwsDailyPkgVerUsrcAdServiceImpl implements DwsService {
         dwBatchMapper.batchInsert(queryStatisticsAll, DwsDailyPkgVerUsrcAdMapper.class);
         // 删除不用的
         List<Long> dbNewAdDelIds = this.getDelIds(dbAdList, queryStatisticsAll, null);
-        dwsDailyPkgVerUsrcAdMapper.deleteBatchIds(dbNewAdDelIds);
+        if (!CollectionUtils.isEmpty(dbNewAdDelIds)) {
+            dwsDailyPkgVerUsrcAdMapper.deleteBatchIds(dbNewAdDelIds);
+        }
         // 发布事件
         applicationEventPublisher.publishEvent(new DwsDailyPkgVerUsrcAdUpdateDwEvent(this, dates));
     }
