+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcOnlineServiceImpl.java
@@ -37,6 +37,8 @@ public class DwsDailyPkgVerUsrcOnlineServiceImpl implements DwsService {
         statisticsList.forEach(DwsDailyPkgVerUsrcOnline::init);
         dwBatchMapper.batchInsert(statisticsList, DwsDailyPkgVerUsrcOnlineMapper.class);
         List<Long> delIds = getDelIds(dbList, statisticsList, null);
-        dwsDailyPkgVerUsrcOnlineMapper.deleteBatchIds(delIds);
+        if (!CollectionUtils.isEmpty(delIds)) {
+            dwsDailyPkgVerUsrcOnlineMapper.deleteBatchIds(delIds);
+        }
     }
 }
