+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcPromotionServiceImpl.java
@@ -41,7 +41,9 @@ public class DwsDailyPkgVerUsrcPromotionServiceImpl implements DwsService {
         dwBatchMapper.batchInsert(statisticsList, DwsDailyPkgVerUsrcPromotionMapper.class);
 
         List<Long> delIds = getDelIds(dbList, statisticsList, null);
-        dwsDailyPkgVerUsrcPromotionMapper.deleteBatchIds(delIds);
+        if (!CollectionUtils.isEmpty(delIds)) {
+            dwsDailyPkgVerUsrcPromotionMapper.deleteBatchIds(delIds);
+        }
 
     }
 }
