+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcRetentionServiceImpl.java
@@ -34,6 +34,8 @@ public class DwsDailyPkgVerUsrcRetentionServiceImpl implements DwsService {
         }
         dwBatchMapper.batchInsert(statisticsList, DwsDailyPkgVerUsrcRetentionMapper.class);
         List<Long> delIds = getDelIds(dbList, statisticsList, null);
-        dwsDailyPkgVerUsrcRetentionMapper.deleteBatchIds(delIds);
+        if (!CollectionUtils.isEmpty(delIds)) {
+            dwsDailyPkgVerUsrcRetentionMapper.deleteBatchIds(delIds);
+        }
     }
 }
