+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcShareServiceImpl.java
@@ -35,6 +35,8 @@ public class DwsDailyPkgVerUsrcShareServiceImpl implements DwsService {
         statisticsList.forEach(DwsDailyPkgVerUsrcShare::init);
         dwBatchMapper.batchInsert(statisticsList, DwsDailyPkgVerUsrcShareMapper.class);
         List<Long> delIds = getDelIds(dbList, statisticsList, null);
-        dwsDailyPkgVerUsrcShareMapper.deleteBatchIds(delIds);
+        if (!CollectionUtils.isEmpty(delIds)) {
+            dwsDailyPkgVerUsrcShareMapper.deleteBatchIds(delIds);
+        }
     }
 }
