+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dws/src/main/java/com/leo/ad/codriver/dws/service/impl/usrc/DwsDailyPkgVerUsrcWithdrawServiceImpl.java
@@ -44,6 +44,9 @@ public class DwsDailyPkgVerUsrcWithdrawServiceImpl implements DwsService {
         }
         dwBatchMapper.batchInsert(statisticsNewList, DwsDailyPkgVerUsrcWithdrawMapper.class);
         delIds = getDelIds(dbList, statisticsNewList, null);
-        dwsDailyPkgVerUsrcWithdrawMapper.deleteBatchIds(delIds);
+
+        if (!CollectionUtils.isEmpty(delIds)) {
+            dwsDailyPkgVerUsrcWithdrawMapper.deleteBatchIds(delIds);
+        }
     }
 }
